+++ src/App.tsx
@@ -1,41 +1,15 @@
-import React from "react";
 import { Layout } from "antd";
-
-const { Header, Sider, Content } = Layout;
-
-const headerStyle: React.CSSProperties = {
-  textAlign: "center",
-  color: "#fff",
-  height: 60,
-  paddingInline: 48,
-  lineHeight: "64px",
-  backgroundColor: "#4096ff",
-};
-
-const contentStyle: React.CSSProperties = {
-  textAlign: "center",
-  minHeight: "calc(100vh - 60px)",
-  color: "#fff",
-  backgroundColor: "#001529",
-};
-
-const siderStyle: React.CSSProperties = {
-  textAlign: "center",
-  lineHeight: "120px",
-  color: "#fff",
-  backgroundColor: "#1677ff",
-};
-
+import AppHeader from "./components/layout/AppHeader";
+import AppSider from "./components/layout/AppSider";
+import AppContent from "./components/layout/AppContent";
 
 function App() {
   return (
     <Layout>
-      <Header style={headerStyle}>Header</Header>
+      <AppHeader/>
       <Layout>
-        <Sider width="25%" style={siderStyle}>
-          Sider
-        </Sider>
-        <Content style={contentStyle}>Content</Content>
+        <AppSider/>
+        <AppContent/>
       </Layout>
     </Layout>
   );
